Add new monorepo package for common code
