+++ packages/common/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['../../.eslintrc.js'],
+  ignorePatterns: ['lib/**']
+}
