+++ packages/common/.prettierrc.js
@@ -1,0 +1,6 @@
+const defaultConfig = require('../../.prettierrc.js')
+const modifiedConfig = {
+  ...defaultConfig
+}
+
+module.exports = modifiedConfig
