+++ packages/common/src/static.ts
@@ -1,0 +1,10 @@
+export const ONION_ADDRESS_REGEX = /^[a-z0-9]{56}$/g
+
+export enum InvitationParams {
+  CODE = 'code'
+}
+
+export enum Site {
+  DOMAIN = 'tryquiet.org',
+  JOIN_PAGE = 'join'
+}
