+++ packages/desktop/src/main/main.ts
@@ -12,7 +12,8 @@ import logger from './logger'
 import { DATA_DIR, DEV_DATA_DIR } from '../shared/static'
 import { fork, ChildProcess } from 'child_process'
 import { getFilesData } from '../utils/functions/fileData'
-import { updateDesktopFile, processInvitationCode, argvInvitationCode, retrieveInvitationCode } from './invitation'
+import { updateDesktopFile, processInvitationCode } from './invitation'
+import { argvInvitationCode, retrieveInvitationCode } from '@quiet/common'
 const ElectronStore = require('electron-store')
 ElectronStore.initRenderer()
