+++ packages/desktop/src/renderer/components/CreateJoinCommunity/PerformCommunityActionComponent.tsx
@@ -15,7 +15,7 @@ import {
   JoinCommunityDictionary
 } from '../CreateJoinCommunity/community.dictionary'
 
-import { parseName, CommunityOwnership, ONION_ADDRESS_REGEX, getInvitationCode } from '@quiet/state-manager'
+import { parseName, CommunityOwnership, getInvitationCode } from '@quiet/state-manager'
 
 import { Controller, useForm } from 'react-hook-form'
 import { TextInput } from '../../forms/components/textInput'
@@ -23,6 +23,7 @@ import { InviteLinkErrors } from '../../forms/fieldsErrors'
 import { IconButton, InputAdornment } from '@mui/material'
 import VisibilityOff from '@mui/icons-material/VisibilityOff'
 import Visibility from '@mui/icons-material/Visibility'
+import { ONION_ADDRESS_REGEX } from '@quiet/common'
 
 const PREFIX = 'PerformCommunityActionComponent'
