+++ packages/desktop/src/renderer/sagas/invitation/handleInvitationCode.saga.ts
@@ -1,7 +1,8 @@
 import { PayloadAction } from '@reduxjs/toolkit'
 import { select, put, delay } from 'typed-redux-saga'
-import { communities, CommunityOwnership, CreateNetworkPayload, ONION_ADDRESS_REGEX } from '@quiet/state-manager'
+import { communities, CommunityOwnership, CreateNetworkPayload } from '@quiet/state-manager'
 import { socketSelectors } from '../socket/socket.selectors'
+import { ONION_ADDRESS_REGEX } from '@quiet/common'
 
 export function* handleInvitationCodeSaga(
     action: PayloadAction<ReturnType<typeof communities.actions.handleInvitationCode>['payload']>
