+++ packages/state-manager/src/index.ts
@@ -85,7 +85,7 @@ export { formatBytes } from './utils/functions/formatBytes/formatBytes'
 
 export { sortPeers } from './utils/functions/sortPeers/sortPeers'
 
-export { getInvitationCode, invitationDeepUrl, invitationShareUrl } from './utils/functions/invitationCode/invitationCode'
+export { getInvitationCode } from './utils/functions/invitationCode/invitationCode'
 
 export type { Socket } from './types'
