+++ tsconfig.json
@@ -7,6 +7,7 @@
         "@quiet/state-manager": ["packages/state-manager/src/index.ts"],
         "@quiet/backend": ["packages/backend/src/index.ts"],
         "@quiet/logger": ["packages/logger/src/index.ts"],
+        "@quiet/common": ["packages/common/src/index.ts"],
         "integration-tests": ["packages/integration-tests/src/index.ts"],
       },
       "module": "commonjs",
